Create `pagination.rs` in pointercrate-core-api
Move the `pagination_response` function here, in anticipation that we
will have more than just a single pagination related function in the
future.
