+++ pointercrate-core-api/src/lib.rs
@@ -3,3 +3,4 @@ pub mod etag;
 pub mod maintenance;
 pub mod query;
 pub mod response;
+pub mod pagination;
