+++ pointercrate-core-api/src/pagination.rs
@@ -1,0 +1,117 @@
+use std::collections::BTreeMap;
+
+use pointercrate_core::pagination::{Pagination, PaginationParameters};
+use rocket::serde::json::Json;
+use serde::Serialize;
+
+use crate::response::Response2;
+
+
+
+pub fn pagination_response<P, T, F>(
+    endpoint: &'static str, mut objects: Vec<T>, paginate: P, min_id: i32, max_id: i32, id_func: F,
+) -> Response2<Json<Vec<T>>>
+where
+    F: Fn(&T) -> i32,
+    P: Pagination,
+    T: Serialize,
+{
+    let parameters = paginate.parameters();
+    // Use a BTreeMap so that we retain insertion order
+    let mut rel = BTreeMap::new();
+
+    rel.insert(
+        "first",
+        paginate.with_parameters(PaginationParameters {
+            before: None,
+            after: Some(min_id - 1),
+            ..parameters
+        }),
+    );
+    rel.insert(
+        "last",
+        paginate.with_parameters(PaginationParameters {
+            before: Some(max_id + 1),
+            after: None,
+            ..parameters
+        }),
+    );
+
+    let limit = parameters.limit as usize;
+    let next_page_exists = objects.len() > limit;
+
+    if !objects.is_empty() {
+        if next_page_exists {
+            objects.pop(); // remove the things from then next page
+        }
+
+        let last_id = id_func(objects.last().unwrap());
+        let first_id = id_func(objects.first().unwrap());
+
+        match (parameters.before, parameters.after) {
+            (None, after) => {
+                // no 'before' value set.
+                // if 'after' is none, we're on the first page, otherwise we have ot generate a 'prev' link
+
+                if next_page_exists {
+                    rel.insert(
+                        "next",
+                        paginate.with_parameters(PaginationParameters {
+                            before: None,
+                            after: Some(last_id),
+                            ..parameters
+                        }),
+                    );
+                }
+
+                if after.is_some() {
+                    rel.insert(
+                        "prev",
+                        paginate.with_parameters(PaginationParameters {
+                            before: Some(first_id),
+                            after: None,
+                            ..parameters
+                        }),
+                    );
+                }
+            },
+            (Some(_), None) => {
+                // A previous page exists. In this case, the page was retrieved using 'ORDER BY ... DESC' so we need to reverse list order!
+                objects.reverse();
+
+                // This means "first" and "last" are actually to opposite of what the variables are named.
+                rel.insert(
+                    "next",
+                    paginate.with_parameters(PaginationParameters {
+                        before: None,
+                        after: Some(first_id),
+                        ..parameters
+                    }),
+                );
+
+                rel.insert(
+                    "prev",
+                    paginate.with_parameters(PaginationParameters {
+                        before: Some(last_id),
+                        after: None,
+                        ..parameters
+                    }),
+                );
+            },
+            (Some(_before), Some(_after)) => {
+                // We interpret this as that all objects _up to 'before'_ are supposed to be paginated.
+                // This means we keep the 'before' value and handle the 'after' value just as above.
+                // tODO: implement
+            },
+        }
+    }
+
+    // Would love to have Iterator::intersperse here
+    let links = rel
+        .into_iter()
+        .map(|(tag, paginate)| format!("<{}?{}>; rel={}", endpoint, serde_urlencoded::to_string(paginate).unwrap(), tag))
+        .collect::<Vec<_>>()
+        .join(",");
+
+    Response2::json(objects).with_header("Links", links)
+}
