+++ pointercrate-demonlist-api/src/endpoints/record.rs
@@ -4,8 +4,8 @@ use pointercrate_core::{audit::AuditLogEntry, error::CoreError, pool::Pointercra
 use pointercrate_core_api::{
     error::Result,
     etag::{Precondition, TaggableExt, Tagged},
+    pagination::pagination_response,
     query::Query,
-    response::pagination_response,
     response::Response2,
 };
 use pointercrate_demonlist::{
